fix: POS checkout 'Invalid argument(s): 0.0' error
- Fix PdfPageFormat height=0 causing ArgumentError in invoice generation (use double.infinity for auto-height)
- Add explicit .toDouble() calls in order items payload for type safety
- Separate print from checkout flow so printing failures don't break the sale
- Better error display with duration and snackPosition

+++ branch_pos/lib/services/supabase_service.dart
@@ -88,8 +88,8 @@ class SupabaseService {
       'product_id': item.productId,
       'product_name': item.name,
       'quantity': item.quantity,
-      'unit_price': item.price,
-      'total_price': item.total,
+      'unit_price': item.price.toDouble(),
+      'total_price': item.total.toDouble(),
     }).toList();
 
     await supabase.from('order_items').insert(orderItems);
